requirements.txt
Flask
gunicorn
python-dotenv
google-generativeai
